+++ src/fvm/fvm_to_histogram.h
@@ -9,7 +9,7 @@
 /*
   This file is part of code_saturne, a general-purpose CFD tool.
 
-  Copyright (C) 1998-2025 EDF S.A.
+  Copyright (C) 1998-2026 EDF S.A.
 
   This program is free software; you can redistribute it and/or modify it under
   the terms of the GNU General Public License as published by the Free Software
@@ -55,7 +55,7 @@ BEGIN_C_DECLS
 typedef enum {
   CS_HISTOGRAM_TXT,  /* .txt file */
   CS_HISTOGRAM_TEX,  /* .tex file */
-  CS_HISTOGRAM_PNG   /* .png file */
+  CS_HISTOGRAM_SVG   /* .svg file */
 } cs_histogram_format_t;
 
 /*----------------------------------------------------------------------------