+++ src/fvm/fvm_writer.cpp
@@ -5,7 +5,7 @@
 /*
   This file is part of code_saturne, a general-purpose CFD tool.
 
-  Copyright (C) 1998-2025 EDF S.A.
+  Copyright (C) 1998-2026 EDF S.A.
 
   This program is free software; you can redistribute it and/or modify it under
   the terms of the GNU General Public License as published by the Free Software
@@ -283,7 +283,7 @@ static fvm_writer_format_t _fvm_writer_format_list[11] = {
 #  else
     0,                                 /* dynamic library flags */
 #  endif
-    nullptr,                              /* dynamic library */
+    nullptr,                           /* dynamic library */
 #if defined(HAVE_CATALYST)
     "fvm_catalyst",                    /* dynamic library name */
     "fvm_to_catalyst_",                /* dynamic library prefix */